Correção na correção da classe card desnecessária e adaptação da classe CreditCard.
