+++ src/Cielo/API30/Ecommerce/Request/UpdateSaleRequest.php
@@ -1,7 +1,6 @@
 <?php
 namespace Cielo\API30\Ecommerce\Request;
 
-use Cielo\API30\Ecommerce\Request\AbstractRequest;
 use Cielo\API30\Environment;
 use Cielo\API30\Merchant;
 use Cielo\API30\Ecommerce\Payment;
